moved kafka setup test code to SenderTest
