+++ src/test/java/com/revature/gambit/GambitTest.java
@@ -1,32 +1,11 @@
 package com.revature.gambit;
 
-import java.util.Map;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.TimeUnit;
-
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.kafka.listener.KafkaMessageListenerContainer;
-import org.springframework.kafka.listener.MessageListener;
-import org.springframework.kafka.listener.config.ContainerProperties;
-import org.springframework.kafka.test.rule.KafkaEmbedded;
-import org.springframework.kafka.test.utils.ContainerTestUtils;
-import org.springframework.kafka.test.utils.KafkaTestUtils;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.revature.gambit.messaging.Sender;
 
 /**
  * Parent test configuration class.
@@ -38,70 +17,7 @@ import com.revature.gambit.messaging.Sender;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql(scripts = "/data-h2.sql", executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
 @Sql(scripts = "/tear-down.sql", executionPhase = ExecutionPhase.AFTER_TEST_METHOD)
-@DirtiesContext
 public class GambitTest {
-
-	private KafkaMessageListenerContainer<String, String> container;
-
-	private BlockingQueue<ConsumerRecord<String, String>> records;
-  
-	private ObjectMapper mapper = new ObjectMapper();
-
-	@ClassRule
-	public static KafkaEmbedded embeddedKafka = 
-		new KafkaEmbedded(1, true, 3);
-
-	@Before
-	public void setUp() throws Exception {
-		// set up the Kafka consumer properties
-		Map<String, Object> consumerProperties =
-				KafkaTestUtils.consumerProps("sender", "false", embeddedKafka);
-
-		// create a Kafka consumer factory
-		DefaultKafkaConsumerFactory<String, String> consumerFactory =
-				new DefaultKafkaConsumerFactory<String, String>(consumerProperties);
-
-		// set the topic that needs to be consumed
-		ContainerProperties containerProperties = new ContainerProperties(
-				"trainer.register.t", "trainer.update.t", "trainer.delete.t");
-
-		// create a Kafka MessageListenerContainer
-		container = new KafkaMessageListenerContainer<>(consumerFactory, containerProperties);
-
-		// create a thread safe queue to store the received message
-		records = new LinkedBlockingQueue<>();
-
-		// setup a Kafka message listener
-		container.setupMessageListener(new MessageListener<String, String>() {
-			@Override
-			public void onMessage(ConsumerRecord<String, String> record) {
-				//logger.debug("test-listener received message=" + record.toString());
-				records.add(record);
-			}
-		});
-
-		// start the container and underlying message listener
-		container.start();
-
-		// wait until the container has the required number of assigned partitions
-		ContainerTestUtils.waitForAssignment(container, embeddedKafka.getPartitionsPerTopic());
-	}
-
-	@After
-	public void tearDown() {
-		// stop the container
-		container.stop();
-	}
-  
-	public Object receive(Class<?> clazz) throws InterruptedException{
-		// Use this to receive object from mock kafka server. It will unmarshalled the json.
-		ConsumerRecord<String, String> received = records.poll(10, TimeUnit.SECONDS);
-		try {
-			return mapper.readValue(received.value(), clazz);
-		} catch (Exception e) {
-			return null;
-		}  
-	}
 	
 	@Test
 	public void run() {}
